spark_job/requirements.txt
wheel
pyspark==2.4.3
awsglue-local==1.0.2